Move bellatrix configuration to be comparable to altair
In order to define Capella configuration, I need to move Bellatrix config into the same kind of structure as Altair, so that I can build on it from there.

Partially addresses #5680
