+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/config/SpecConfigBellatrix.java
@@ -13,66 +13,15 @@
 
 package tech.pegasys.teku.spec.config;
 
-import java.util.Objects;
 import java.util.Optional;
 import org.apache.tuweni.bytes.Bytes32;
 import org.apache.tuweni.units.bigints.UInt256;
 import tech.pegasys.teku.infrastructure.bytes.Bytes4;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public class SpecConfigBellatrix extends DelegatingSpecConfigAltair {
+public interface SpecConfigBellatrix extends SpecConfigAltair {
 
-  // Fork
-  private final Bytes4 bellatrixForkVersion;
-  private final UInt64 bellatrixForkEpoch;
-  private final UInt64 inactivityPenaltyQuotientBellatrix;
-  private final int minSlashingPenaltyQuotientBellatrix;
-  private final int proportionalSlashingMultiplierBellatrix;
-  private final int maxBytesPerTransaction;
-  private final int maxTransactionsPerPayload;
-  private final int bytesPerLogsBloom;
-  private final int maxExtraDataBytes;
-
-  // Transition
-  private final UInt256 terminalTotalDifficulty;
-  private final Bytes32 terminalBlockHash;
-  private final UInt64 terminalBlockHashActivationEpoch;
-
-  // Optimistic Sync
-  private final int safeSlotsToImportOptimistically;
-
-  public SpecConfigBellatrix(
-      final SpecConfigAltair specConfig,
-      final Bytes4 bellatrixForkVersion,
-      final UInt64 bellatrixForkEpoch,
-      final UInt64 inactivityPenaltyQuotientBellatrix,
-      final int minSlashingPenaltyQuotientBellatrix,
-      final int proportionalSlashingMultiplierBellatrix,
-      final int maxBytesPerTransaction,
-      final int maxTransactionsPerPayload,
-      final int bytesPerLogsBloom,
-      final int maxExtraDataBytes,
-      final UInt256 terminalTotalDifficulty,
-      final Bytes32 terminalBlockHash,
-      final UInt64 terminalBlockHashActivationEpoch,
-      final int safeSlotsToImportOptimistically) {
-    super(specConfig);
-    this.bellatrixForkVersion = bellatrixForkVersion;
-    this.bellatrixForkEpoch = bellatrixForkEpoch;
-    this.inactivityPenaltyQuotientBellatrix = inactivityPenaltyQuotientBellatrix;
-    this.minSlashingPenaltyQuotientBellatrix = minSlashingPenaltyQuotientBellatrix;
-    this.proportionalSlashingMultiplierBellatrix = proportionalSlashingMultiplierBellatrix;
-    this.maxBytesPerTransaction = maxBytesPerTransaction;
-    this.maxTransactionsPerPayload = maxTransactionsPerPayload;
-    this.bytesPerLogsBloom = bytesPerLogsBloom;
-    this.maxExtraDataBytes = maxExtraDataBytes;
-    this.terminalTotalDifficulty = terminalTotalDifficulty;
-    this.terminalBlockHash = terminalBlockHash;
-    this.terminalBlockHashActivationEpoch = terminalBlockHashActivationEpoch;
-    this.safeSlotsToImportOptimistically = safeSlotsToImportOptimistically;
-  }
-
-  public static SpecConfigBellatrix required(final SpecConfig specConfig) {
+  static SpecConfigBellatrix required(SpecConfig specConfig) {
     return specConfig
         .toVersionBellatrix()
         .orElseThrow(
@@ -82,103 +31,32 @@ public class SpecConfigBellatrix extends DelegatingSpecConfigAltair {
                         + specConfig.getClass().getSimpleName()));
   }
 
-  public Bytes4 getBellatrixForkVersion() {
-    return bellatrixForkVersion;
-  }
-
-  public UInt64 getBellatrixForkEpoch() {
-    return bellatrixForkEpoch;
-  }
-
-  public UInt64 getInactivityPenaltyQuotientBellatrix() {
-    return inactivityPenaltyQuotientBellatrix;
-  }
-
-  public int getMinSlashingPenaltyQuotientBellatrix() {
-    return minSlashingPenaltyQuotientBellatrix;
-  }
-
-  public int getProportionalSlashingMultiplierBellatrix() {
-    return proportionalSlashingMultiplierBellatrix;
-  }
-
-  public int getMaxBytesPerTransaction() {
-    return maxBytesPerTransaction;
-  }
-
-  public int getMaxTransactionsPerPayload() {
-    return maxTransactionsPerPayload;
-  }
-
-  public int getBytesPerLogsBloom() {
-    return bytesPerLogsBloom;
-  }
-
-  public int getMaxExtraDataBytes() {
-    return maxExtraDataBytes;
-  }
-
-  public UInt256 getTerminalTotalDifficulty() {
-    return terminalTotalDifficulty;
-  }
-
-  public Bytes32 getTerminalBlockHash() {
-    return terminalBlockHash;
-  }
-
-  public UInt64 getTerminalBlockHashActivationEpoch() {
-    return terminalBlockHashActivationEpoch;
-  }
-
-  public int getSafeSlotsToImportOptimistically() {
-    return safeSlotsToImportOptimistically;
-  }
-
   @Override
-  public Optional<SpecConfigBellatrix> toVersionBellatrix() {
-    return Optional.of(this);
-  }
+  Optional<SpecConfigBellatrix> toVersionBellatrix();
 
-  @Override
-  public boolean equals(final Object o) {
-    if (this == o) {
-      return true;
-    }
-    if (o == null || getClass() != o.getClass()) {
-      return false;
-    }
-    final SpecConfigBellatrix that = (SpecConfigBellatrix) o;
-    return Objects.equals(specConfig, that.specConfig)
-        && minSlashingPenaltyQuotientBellatrix == that.minSlashingPenaltyQuotientBellatrix
-        && proportionalSlashingMultiplierBellatrix == that.proportionalSlashingMultiplierBellatrix
-        && maxBytesPerTransaction == that.maxBytesPerTransaction
-        && maxTransactionsPerPayload == that.maxTransactionsPerPayload
-        && bytesPerLogsBloom == that.bytesPerLogsBloom
-        && maxExtraDataBytes == that.maxExtraDataBytes
-        && Objects.equals(bellatrixForkVersion, that.bellatrixForkVersion)
-        && Objects.equals(bellatrixForkEpoch, that.bellatrixForkEpoch)
-        && Objects.equals(
-            inactivityPenaltyQuotientBellatrix, that.inactivityPenaltyQuotientBellatrix)
-        && Objects.equals(terminalTotalDifficulty, that.terminalTotalDifficulty)
-        && Objects.equals(terminalBlockHash, that.terminalBlockHash)
-        && Objects.equals(terminalBlockHashActivationEpoch, that.terminalBlockHashActivationEpoch);
-  }
+  Bytes4 getBellatrixForkVersion();
 
-  @Override
-  public int hashCode() {
-    return Objects.hash(
-        specConfig,
-        bellatrixForkVersion,
-        bellatrixForkEpoch,
-        inactivityPenaltyQuotientBellatrix,
-        minSlashingPenaltyQuotientBellatrix,
-        proportionalSlashingMultiplierBellatrix,
-        maxBytesPerTransaction,
-        maxTransactionsPerPayload,
-        bytesPerLogsBloom,
-        maxExtraDataBytes,
-        terminalTotalDifficulty,
-        terminalBlockHash,
-        terminalBlockHashActivationEpoch);
-  }
+  UInt64 getBellatrixForkEpoch();
+
+  UInt64 getInactivityPenaltyQuotientBellatrix();
+
+  int getMinSlashingPenaltyQuotientBellatrix();
+
+  int getProportionalSlashingMultiplierBellatrix();
+
+  int getMaxBytesPerTransaction();
+
+  int getMaxTransactionsPerPayload();
+
+  int getBytesPerLogsBloom();
+
+  int getMaxExtraDataBytes();
+
+  UInt64 getTerminalBlockHashActivationEpoch();
+
+  Bytes32 getTerminalBlockHash();
+
+  UInt256 getTerminalTotalDifficulty();
+
+  int getSafeSlotsToImportOptimistically();
 }
